musebackendserver/requirements.txt
asgiref==3.2.10
attrs==21.2.0
Automat==20.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
Django==3.1.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
gunicorn==20.1.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
itemadapter==0.3.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.3
mysql-connector-python==8.0.21
parsel==1.6.0
priority==1.3.0
Protego==0.1.16
protobuf==3.17.3
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
python-scrapyd-api==2.1.2
pytz==2021.1
queuelib==1.6.1
requests==2.26.0
Scrapy==2.3.0
scrapyd==1.2.1
scrapyd-client==1.2.0a1
service-identity==21.1.0
six==1.16.0
spotipy==2.18.0
sqlparse==0.4.1
Twisted==21.2.0
typing-extensions==3.10.0.0
urllib3==1.26.6
w3lib==1.22.0
zope.interface==5.4.0
